+++ record_platform_interface/lib/record_platform_interface.dart
@@ -1,2 +1,2 @@
-export 'package:record_platform_interface/src/record_platform_interface.dart';
-export 'package:record_platform_interface/src/types/types.dart';
+export 'src/record_platform_interface.dart';
+export 'src/types/types.dart';