+++ record_platform_interface/lib/src/record_platform.dart
@@ -1,0 +1,7 @@
+import 'record_method_channel_mixin.dart';
+import 'record_platform_interface.dart';
+
+class RecordPlatformImpl extends RecordPlatform
+    with RecordMethodChannel, RecordEventChannel {
+  RecordPlatformImpl();
+}